+++ src/FsCodec/Union.fs
@@ -26,10 +26,11 @@ module Info =
         let getCase value = cases[getTag value]
         { cases = cases; getCase = getCase })
     let tryFindCaseWithName u (predicate: string -> bool): CaseInfo option = u.cases |> Array.tryFind (fun c -> predicate c.name)
-    let private caseValues: Type -> obj[] = memoize (fun t -> (get t).cases |> Array.map (fun c -> c.construct Array.empty))
+    let caseValues<'t>: 't[] = (get typeof<'t>).cases |> Array.map (fun c -> c.construct Array.empty :?> 't)
+    let caseValuesT: Type -> obj[] = memoize (fun t -> (get t).cases |> Array.map (fun c -> c.construct Array.empty))
     let tryFindCaseValueWithName (t: Type): (string -> bool) -> obj option =
         let u = get t
-        let caseValue = let values = caseValues t in fun i -> values[i]
+        let caseValue = let values = caseValuesT t in fun i -> values[i]
         fun predicate -> u.cases |> Array.tryFindIndex (fun c -> predicate c.name) |> Option.map caseValue
 
 /// Determines whether the type is a Union